requirements.txt
playwright
requests
pandas
bs4
python-dotenv
google-api-python-client
google-auth